+++ sdl3-main/ChangeLog.md
@@ -6,6 +6,14 @@
       It still exists for compatibility, but it's hidden in the docs.
     - Pass arguments to main on `std` (`!std` already did)
 
+- 0.5.4:
+    - Fix accidental dependency on Rust 1.84
+
+- 0.5.3:
+    - Only use the alloc-less optimization of `run_async_on_main_thread` if the size of
+      the callback is zero, because that's the only way to guarantee we won't copy
+      uninit bytes, which would be unsound
+
 - 0.5.2:
     - Fix more unsoundness in `run_async_on_main_thread`